refactor: move flow_api module out of tools directory and format code
